server/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Coders for Causes"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
psycopg = {extras = ["binary", "pool"], version = "^3.2.4"}
freezegun = "^1.5.1"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
django-extensions = "^3.2.3"
djangorestframework-simplejwt = "^5.4.0"
drf-spectacular = "^0.28.0"
django-filter = "^24.3"
pillow = "^11.1.0"
django-unfold = "^0.50.0"
django-import-export = "^4.3.7"
python-docx = "^1.2.0"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
flake8 = "6.1.0"
flake8-django = "^1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
